run_sotest: Add poll app for polling without creation
