+++ pkgs/run-sotest/src/run_sotest/__main__.py
@@ -9,10 +9,10 @@ import argparse
 import os
 import requests
 import sys
-import time
+
+from .poll_test_run import poll_test_run
 
 HTTP_STATUS_OK = 200
-REQUEST_HEADERS = {"Accept": "application/json"}
 
 
 def parse_cmdline_args():
@@ -79,35 +79,6 @@ def create_test_run(args):
         sys.exit("Testrun Creation failed: {}".format(r.text))
 
     return r.text
-
-
-def query_test_run(sotest_url, testrun_id):
-    """Queries the status of a test run"""
-    url = "{}/test_runs/{}/status".format(sotest_url, testrun_id)
-    r = requests.get(url, headers=REQUEST_HEADERS)
-
-    if r.status_code != HTTP_STATUS_OK:
-        sys.exit("Testrun query failed: {}".format(r.text))
-
-    return r.json()
-
-
-def poll_test_run(sotest_url, testrun_id):
-    """Queries the status of a test run until it is no longer running and handles the result"""
-    while True:
-        result = query_test_run(sotest_url, testrun_id)
-        if result != "unfinished":
-            break
-        time.sleep(10)
-
-    if result == "success":
-        print("Test successful")
-    elif result == "fail":
-        sys.exit("Test failed")
-    elif result == "disabled":
-        sys.exit("Test has been aborted")
-    else:
-        sys.exit("Unexpected response: {}".format(result))
 
 
 def main():
